requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
docker==7.1.0
email_validator==2.1.1
fastapi==0.111.0
fastapi-cli==0.0.4
frozenlist==1.4.1
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.4
magic-filter==1.0.12
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
minio==7.2.7
multidict==6.0.5
orjson==3.10.4
packaging==24.1
pluggy==1.5.0
psycopg2-binary==2.9.9
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
pytest==8.2.2
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.30
starlette==0.37.2
typer==0.12.3
typing_extensions==4.11.0
ujson==5.10.0
urllib3==1.26.18
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
yarl==1.9.4
